# Optimize nonbonded tests