--- a/.pre-commit-config.yaml
+++ b/.pre-commit-config.yaml
@@ -16,6 +16,7 @@ repos:
     rev: 21.10b0
     hooks:
     -   id: black
+        additional_dependencies: ['click==8.0.4'] # https://github.com/psf/black/issues/2964#issuecomment-1080974737
 
 -   repo: https://github.com/pycqa/flake8
     rev: 4.0.1