--- a/tests/nonbonded/test_consistency.py
+++ b/tests/nonbonded/test_consistency.py
@@ -30,9 +30,7 @@ def filter_valid_exclusions(
 @pytest.mark.parametrize("precision", [np.float64, np.float32])
 @pytest.mark.parametrize("num_atoms,num_atoms_ligand", [(33, 1), (4080, 1050)])
 @pytest.mark.parametrize("interpolated", [False, True])
-@pytest.mark.parametrize("disable_hilbert_sort", [False, True])
 def test_nonbonded_consistency(
-    disable_hilbert_sort,
     interpolated,
     num_atoms_ligand,
     num_atoms,
@@ -65,9 +63,6 @@ def test_nonbonded_consistency(
         exclusion_idxs, exclusion_scales, lambda_plane_idxs, lambda_offset_idxs, beta, cutoff
     ).unbound_impl(precision)
 
-    if disable_hilbert_sort:
-        ref_impl.disable_hilbert_sort()
-
     def make_allpairs_potential(atom_idxs):
         make_potential = NonbondedAllPairsInterpolated if interpolated else NonbondedAllPairs
         return make_potential(
@@ -76,7 +71,6 @@ def test_nonbonded_consistency(
             beta,
             cutoff,
             atom_idxs,
-            disable_hilbert_sort=disable_hilbert_sort,
         )
 
     def make_ixngroup_potential(ligand_idxs):
@@ -87,7 +81,6 @@ def test_nonbonded_consistency(
             lambda_offset_idxs,
             beta,
             cutoff,
-            disable_hilbert_sort=disable_hilbert_sort,
         )
 
     def make_pairlist_potential(exclusion_idxs, exclusion_scales):
@@ -103,41 +96,55 @@ def test_nonbonded_consistency(
         ]
     ).unbound_impl(precision)
 
-    for lam in [0.0, 0.1]:
-        du_dx_ref, du_dp_ref, du_dl_ref, u_ref = ref_impl.execute(conf, params, example_box, lam)
-        du_dx_test, du_dp_test, du_dl_test, u_test = test_impl.execute(conf, params, example_box, lam)
+    def test():
+        for lam in [0.0, 0.1]:
+            du_dx_ref, du_dp_ref, du_dl_ref, u_ref = ref_impl.execute(conf, params, example_box, lam)
+            du_dx_test, du_dp_test, du_dl_test, u_test = test_impl.execute(conf, params, example_box, lam)
 
-        np.testing.assert_array_equal(du_dx_test, du_dx_ref)
+            np.testing.assert_array_equal(du_dx_test, du_dx_ref)
 
-        if interpolated:
-            # NOTE: bitwise equivalence is not currently possible for the
-            # interpolated case. To see this, note that the interpolated
-            # energy is given by
-            #
-            #   u(p0, p1) = (1 - lam) * F(p0) + lam * F(p1)
-            #
-            # In particular,
-            #
-            #   du_dp1 = lam * f(p1)
-            #
-            # where f(p) = F'(p). The reference potential effectively sums
-            # over interactions before multiplication by \lambda
-            #
-            #   du_dp1_ref = lam * fixed_sum(f_host(p1), f_ligand(p1), f_host_ligand(p1))
-            #
-            # while the test potential (because it's implemented using
-            # SummedPotential), effectively distributes multiplication by
-            # \lambda into the sum
-            #
-            #   du_dp1_test = fixed_sum(lam * f_host(p1), lam * f_ligand(p1), lam * f_host_ligand(p1))
-            #
-            # Since `c * fixed_sum(x, y)` is not bitwise equivalent to
-            # `fixed_sum(c * x, c * y)` in general, the reference and test
-            # du_dps are not guaranteed to be bitwise equivalent in the
-            # interpolated case.
-            np.testing.assert_allclose(du_dp_test, du_dp_ref, rtol=1e-8, atol=1e-8)
-        else:
-            np.testing.assert_array_equal(du_dp_test, du_dp_ref)
+            if interpolated:
+                # NOTE: bitwise equivalence is not currently possible for the
+                # interpolated case. To see this, note that the interpolated
+                # energy is given by
+                #
+                #   u(p0, p1) = (1 - lam) * F(p0) + lam * F(p1)
+                #
+                # In particular,
+                #
+                #   du_dp1 = lam * f(p1)
+                #
+                # where f(p) = F'(p). The reference potential effectively sums
+                # over interactions before multiplication by \lambda
+                #
+                #   du_dp1_ref = lam * fixed_sum(f_host(p1), f_ligand(p1), f_host_ligand(p1))
+                #
+                # while the test potential (because it's implemented using
+                # SummedPotential), effectively distributes multiplication by
+                # \lambda into the sum
+                #
+                #   du_dp1_test = fixed_sum(lam * f_host(p1), lam * f_ligand(p1), lam * f_host_ligand(p1))
+                #
+                # Since `c * fixed_sum(x, y)` is not bitwise equivalent to
+                # `fixed_sum(c * x, c * y)` in general, the reference and test
+                # du_dps are not guaranteed to be bitwise equivalent in the
+                # interpolated case.
+                np.testing.assert_allclose(du_dp_test, du_dp_ref, rtol=1e-8, atol=1e-8)
+            else:
+                np.testing.assert_array_equal(du_dp_test, du_dp_ref)
 
-        np.testing.assert_array_equal(du_dl_test, du_dl_ref)
-        assert u_test == u_ref
+            np.testing.assert_array_equal(du_dl_test, du_dl_ref)
+            assert u_test == u_ref
+
+    test()
+
+    # Test with hilbert sorting disabled
+    ref_impl.disable_hilbert_sort()
+
+    # NonbondedAllPairs and NonbondedInteractionGroup have a
+    # disable_hilbert_sort method; NonbondedPairList doesn't
+    for impl in test_impl.get_potentials():
+        if hasattr(impl, "disable_hilbert_sort"):
+            impl.disable_hilbert_sort()
+
+    test()