New tests added as part of the nonbonded potential split take up 4 spots in the top 10 longest-running tests. This PR applies several optimizations to make these run faster without reducing coverage:

- Move loop over lambdas from `pytest.mark.parameterize` decorator to a loop inside the test function. This prevents reconstructing the test potentials for every lambda value. (Constructing the unbound impls is by far the most costly step.)
- Similarly, move loop over `disable_hilbert` into the test function (also remove method overrides in `NonbondedAllPairs` and `NonbondedInteractionGroup` that aren't used anymore)

These changes cut total test time down by about 5 minutes.

<details>

### Top 10 test durations

#### Before

| test_suite     | path                                                | name                                                                 |   test_count |   duration |   fraction_total_duration |
|:---------------|:----------------------------------------------------|:---------------------------------------------------------------------|-------------:|-----------:|--------------------------:|
| unit_tests     | tests/nonbonded/test_consistency.py                 | test_nonbonded_consistency                                           |           32 |     227.17 |                 0.134013  |
| memcheck_tests | tests/nonbonded/test_nonbonded_all_pairs.py         | test_nonbonded_all_pairs_correctness                                 |           48 |      84.21 |                 0.0496773 |
| unit_tests     | tests/test_mtm.py                                   | test_optimized_MTM                                                   |            1 |      82.89 |                 0.0488986 |
| unit_tests     | tests/test_docking.py                               | TestDocking::test_rigorous_work                                      |            1 |      77.02 |                 0.0454358 |
| memcheck_tests | tests/test_nonbonded.py                             | TestNonbondedDHFR::test_correctness                                  |            1 |      61.93 |                 0.0365339 |
| unit_tests     | tests/test_free_energy.py                           | test_relative_free_energy                                            |            1 |      55.99 |                 0.0330297 |
| memcheck_tests | tests/nonbonded/test_nonbonded_interaction_group.py | test_nonbonded_interaction_group_consistency_allpairs_lambda_planes  |           24 |      55.09 |                 0.0324988 |
| memcheck_tests | tests/nonbonded/test_nonbonded_interaction_group.py | test_nonbonded_interaction_group_consistency_allpairs_constant_shift |           16 |      51.26 |                 0.0302394 |
| memcheck_tests | tests/test_nonbonded.py                             | TestNonbonded::test_nonbonded                                        |            1 |      49.57 |                 0.0292424 |
| unit_tests     | tests/test_endpoint_correction.py                   | test_endpoint_correction                                             |            1 |      49.56 |                 0.0292365 |

#### After

| test_suite     | path                                                | name                                                                |   test_count |   duration |   fraction_total_duration |
|:---------------|:----------------------------------------------------|:--------------------------------------------------------------------|-------------:|-----------:|--------------------------:|
| memcheck_tests | tests/nonbonded/test_nonbonded_all_pairs.py         | test_nonbonded_all_pairs_correctness                                |           48 |      82.33 |                 0.0561164 |
| unit_tests     | tests/test_mtm.py                                   | test_optimized_MTM                                                  |            1 |      81.06 |                 0.0552507 |
| unit_tests     | tests/test_docking.py                               | TestDocking::test_rigorous_work                                     |            1 |      76.93 |                 0.0524357 |
| memcheck_tests | tests/test_nonbonded.py                             | TestNonbondedDHFR::test_correctness                                 |            1 |      60.29 |                 0.0410938 |
| unit_tests     | tests/nonbonded/test_consistency.py                 | test_nonbonded_consistency                                          |            8 |      56.85 |                 0.0387491 |
| unit_tests     | tests/test_free_energy.py                           | test_relative_free_energy                                           |            1 |      55.16 |                 0.0375972 |
| memcheck_tests | tests/nonbonded/test_nonbonded_interaction_group.py | test_nonbonded_interaction_group_consistency_allpairs_lambda_planes |           24 |      53.1  |                 0.0361931 |
| memcheck_tests | tests/test_nonbonded.py                             | TestNonbonded::test_nonbonded                                       |            1 |      50.03 |                 0.0341006 |
| unit_tests     | tests/test_endpoint_correction.py                   | test_endpoint_correction                                            |            1 |      49.33 |                 0.0336235 |
| unit_tests     | tests/test_free_energy.py                           | test_absolute_free_energy                                           |            1 |      42.72 |                 0.0291181 |

</details>